+++ gpkit/constraints/set.py
@@ -15,6 +15,7 @@ def _sort_by_name_and_idx(var):
 class ConstraintSet(list):
     "Recursive container for ConstraintSets and Inequalities"
     varkeys = None
+    unique_varkeys = frozenset()
 
     def __init__(self, constraints, substitutions=None):
         if isinstance(constraints, ConstraintSet):
@@ -24,7 +25,6 @@ class ConstraintSet(list):
 
         # initializations for attributes used elsewhere
         self.posymap = []
-        self.unused_variables = None
         self.numpy_bools = False
 
         # get substitutions and convert all members to ConstraintSets
@@ -53,6 +53,8 @@ class ConstraintSet(list):
             return list.__getitem__(self, key)
         else:
             variables = self.variables_byname(key)
+            if not variables:
+                raise KeyError(key)
             if variables[0].key.veckey:
                 # maybe it's all one vector variable!
                 from ..nomials import NomialArray
@@ -166,14 +168,20 @@ class ConstraintSet(list):
         "Substitutes in place."
         for constraint in self:
             constraint.subinplace(subs)
-        if self.unused_variables is not None:
-            unused_vars = []
-            for var in self.unused_variables:
-                if var.key in subs:
-                    unused_vars.append(subs[var.key])
-                else:
-                    unused_vars.append(var.key)
-            self.unused_variables = unused_vars
+        for key in subs:
+            if key not in self.substitutions:
+                continue
+            if hasattr(subs[key], "key"):
+                self.substitutions[subs[key]] = self.substitutions[key]
+                del self.substitutions[key]
+            else:
+                raise ValueError("the substitution {%s: %s} is invalidated"
+                                 " by the subinplace {%s: %s}, because"
+                                 " %s does not have a `key` attribute"
+                                 % (key, self.substitutions[key],
+                                    key, subs[key], subs[key]))
+        self.unique_varkeys = frozenset(subs[vk] if vk in subs else vk
+                                        for vk in self.unique_varkeys)
         self.reset_varkeys()
 
     def reset_varkeys(self, init_dict=None):
@@ -184,8 +192,7 @@ class ConstraintSet(list):
         for constraint in self:
             if hasattr(constraint, "varkeys"):
                 varkeys.update(constraint.varkeys)
-        if self.unused_variables is not None:
-            varkeys.update(self.unused_variables)
+        varkeys.update(self.unique_varkeys)
         self.varkeys = varkeys
         self.substitutions.varkeys = varkeys
 
@@ -229,12 +236,12 @@ class ConstraintSet(list):
             offset += n_posys
         return var_senss
 
-    def as_gpconstr(self, x0):
+    def as_gpconstr(self, x0, substitutions=None):
         """Returns GPConstraint approximating this constraint at x0
 
         When x0 is none, may return a default guess."""
-        gpconstrs = [constr.as_gpconstr(x0) for constr in self]
-        return ConstraintSet(gpconstrs, self.substitutions)
+        gpconstrs = [constr.as_gpconstr(x0, substitutions) for constr in self]
+        return ConstraintSet(gpconstrs, substitutions)
 
     def process_result(self, result):
         """Does arbitrary computation / manipulation of a program's result
@@ -252,6 +259,14 @@ class ConstraintSet(list):
         for constraint in self:
             if hasattr(constraint, "process_result"):
                 constraint.process_result(result)
+        for v in self.unique_varkeys:
+            if not v.evalfn or v in result["variables"]:
+                continue
+            if v.veckey:
+                v = v.veckey
+            val = v.evalfn(result["variables"])
+            result["freevariables"][v] = val
+            result["variables"][v] = val
 
 
 def raise_badelement(cns, i, constraint):