+++ gpkit/nomials/substitution.py
@@ -14,6 +14,7 @@ from .. import DimensionalityError
 def parse_subs(varkeys, substitutions):
     "Seperates subs into constants, sweeps linkedsweeps actually present."
     constants, sweep, linkedsweep = {}, {}, {}
+    varkeys.update_keymap()
     if hasattr(substitutions, "keymap"):
         for var in varkeys.keymap:
             if dict.__contains__(substitutions, var):
@@ -61,15 +62,16 @@ def append_sub(sub, keys, constants, sweep, linkedsweep):
                                  " variable %s of shape %s." %
                                  (sub.shape, key.str_without("model"),
                                   key.shape))
-        if not sweepsub:
+
+        if hasattr(value, "__call__") and not hasattr(value, "key"):
+            linkedsweep[key] = value
+        elif sweepsub:
+            sweep[key] = value
+        else:
             try:
                 assert np.isnan(value)
             except (AssertionError, TypeError, ValueError):
                 constants[key] = value
-        elif not hasattr(value, "__call__"):
-            sweep[key] = value
-        else:
-            linkedsweep[key] = value
 
 
 def substitution(nomial, substitutions):
@@ -147,7 +149,7 @@ def substitution(nomial, substitutions):
                 #                          " '%s' into '%s' of units '%s'." %
                 #                          (sub, var, var.units.units))
                 if sub != 0:
-                    mag(cs_)[i] *= sub**x
+                    mag(cs_)[i] *= float(sub)**x
                 elif x > 0:  # HACK to prevent RuntimeWarnings
                     mag(cs_)[i] = 0
                 elif x < 0: