+++ gpkit/tests/t_solution_array.py
@@ -1,6 +1,5 @@
 """Tests for SolutionArray class"""
 import unittest
-import time
 import numpy as np
 from gpkit import Variable, VectorVariable, Model, SignomialsEnabled
 import gpkit
@@ -36,22 +35,6 @@ class TestSolutionArray(unittest.TestCase):
         self.assertEqual(type(solx), np.ndarray)
         self.assertEqual(solx.shape, (n,))
         self.assertTrue((abs(solx - 2.5*np.ones(n)) < 1e-7).all())
-
-    def test_call_time(self):
-        N = 20
-        x = VectorVariable(N, 'x', 'm')
-        y = VectorVariable(N, 'y', 'm')
-        z1 = VectorVariable(N, 'z1', 'm')
-        z2 = VectorVariable(N, 'z2', 'm')
-        z3 = VectorVariable(N, 'z3', 'm')
-        z4 = VectorVariable(N, 'z4', 'm')
-        L = Variable('L', 5, 'm')
-        prob = Model(sum(x),
-                     [x >= y, y >= z1, z1 >= z2, z2 >= z3, z3 >= z4, z4 >= L])
-        sol = prob.solve(verbosity=0)
-        t1 = time.time()
-        _ = sol(z1)
-        self.assertLess(time.time() - t1, 0.05)
 
     def test_subinto(self):
         Nsweep = 20
@@ -92,6 +75,21 @@ class TestSolutionArray(unittest.TestCase):
         self.assertFalse(
             "1000N" in
             sol.table().replace(" ", "").replace("[", "").replace("]", ""))
+
+    def test_key_options(self):
+        # issue 993
+        x = Variable("x")
+        y = Variable("y")
+        with SignomialsEnabled():
+            m = Model(y, [y + 6*x >= 13 + x**2])
+        msol = m.localsolve(verbosity=0)
+        spsol = m.sp().localsolve(verbosity=0)  # pylint: disable=no-member
+        gpsol = m.program.gps[-1].solve(verbosity=0)
+        self.assertEqual(msol(x), msol("x"))
+        self.assertEqual(spsol(x), spsol("x"))
+        self.assertEqual(gpsol(x), gpsol("x"))
+        self.assertEqual(msol(x), spsol(x))
+        self.assertEqual(msol(x), gpsol(x))
 
 
 class TestResultsTable(unittest.TestCase):